fix(db): rendre les migrations rejouables de zero
`migration:fresh` ne passait pas : aucun environnement neuf ne pouvait
etre provisionne depuis le depot — ni test, ni CI, ni un poste qui arrive.
Personne ne l'avait vu, tout le monde travaillant sur la meme base.

Deux tables n'avaient pas de migration : `countries` — la sienne etait
entierement commentee — et `payment_methods`. Elles sont ecrites d'apres
la structure reelle.

Quatre cles etrangeres declaraient un type incompatible avec leur cible,
une visait une table creee plus tard, une portait une faute de frappe
(`wallet`). Une migration importait des modeles deplaces depuis, ce qui
la cassait retroactivement : reecrite en SQL, une migration passee doit
rester rejouable quel que soit le code.

Les colonnes ajoutees, supprimees ou renommees a la main sont reportees
dans la migration de leur table, pas dans un fichier de rattrapage : ce
sont des reajustements entre le depot et la base, pas de nouvelles
etapes. `wallets.users_id`, `admins.role`, `payments.callback_status` et
`transactions.services_id` partent ; `qrcode_token`, `otps.target`,
`transactions.direction`, les cinq colonnes de diagnostic de `payments`
et six autres entrent.

Dix migrations de tables supprimees en base sont retirees.

Le referentiel des pays manquait : sans lui, aucune inscription ne passe.
Ses 239 lignes sont seedees en conservant les identifiants, qu'un
utilisateur porte. Quatre seeders importaient des modeles deplaces.

`charset: utf8mb4` est retire de la connexion : Knex l'injecte dans
`DEFAULT CHARACTER SET`, ce qui fait prendre a MariaDB 11.4 la collation
par defaut du jeu et ecrase celle du serveur. C'est le serveur qui decide.

Resultat : 91 migrations, 43 tables, schema identique a la production.

+++ database/migrations/17298456241_create_users_table.ts
@@ -6,12 +6,13 @@ export default class extends BaseSchema {
   async up() {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
+      // `integer` et non `bigInteger` : `countries.id` est un `increments()`, donc un `int
+      // unsigned`. Une clé étrangère entre types différents est refusée.
       table
-        .bigInteger('country_id')
+        .integer('country_id')
         .unsigned()
         .nullable()
         .references('id')
-
         .inTable('countries')
         .onDelete('SET NULL')
         .onUpdate('CASCADE')
